+++ src/errors/no_renderings.html
@@ -5,7 +5,7 @@
     </head>
     <body>
         <p>
-            IMAGE is unable to render this graphic. If you would like to let us know about this, please send a link to this graphic and a short description of the problem via email: <a href = "mailto: image@cim.mcgill.ca"> image@cim.mcgill.ca</a> or report bugs using <a id="feedback-a" href="../info/feedback.html" target="_blank">our feedback form</a>.
+            IMAGE is unable to render this graphic. If you would like to let us know about this, please send a link to this graphic and a short description of the problem via email: <a href = "mailto: image@cim.mcgill.ca"> image@cim.mcgill.ca</a> or report bugs using <a id="feedback-a" href="../feedback/feedback.html" target="_blank">our feedback form</a>.
             Please ensure that the graphic you are sending contains no personal information.
         </p>
         <button id="closing-button" type="button" > Close Window </button>
