+++ src/info/info.html
@@ -9,7 +9,7 @@
         </header>
         <main id="renderings-container"></main>
         <footer class="container">
-            <p>How did you like these interpretations? We love getting feedback and bug reports using <a id="feedback-a" href="./feedback.html" target="_blank">our feedback form</a>. For anything else, you can also email us at <a href = "mailto:image@cim.mcgill.ca">image@cim.mcgill.ca</a>. In either case, please make sure you do not send any personal information.</p>
+            <p>How did you like these interpretations? We love getting feedback and bug reports using <a id="feedback-a" href="../feedback/feedback.html" target="_blank">our feedback form</a>. For anything else, you can also email us at <a href = "mailto:image@cim.mcgill.ca">image@cim.mcgill.ca</a>. In either case, please make sure you do not send any personal information.</p>
         </footer>
         <script type="module" src="./info.ts"></script>
     </body>
