+++ src/progressBar/progressBar.ts
@@ -17,7 +17,7 @@ var bar = new ProgressBar.Line('#progress', {
 
 bar.animate(1);
 
-var audio = new Audio('../progressBar/earcon_server_communication_IMAGE_processing-data.mp3');
+var audio = new Audio('./audio-files/earcon_server_communication_IMAGE_processing-data.mp3');
 setInterval(function(){
 	audio.play();
 }, 3000);
